fix: fix vm version select
