+++ script/src/verify.rs
@@ -536,6 +536,18 @@ impl<DL: CellDataProvider + HeaderProvider + ExtensionProvider + Send + Sync + C
         }
     }
 
+    fn is_vm_version_1_and_syscalls_2_enabled(&self) -> bool {
+        // If the proposal window is allowed to prejudge on the vm version,
+        // it will cause proposal tx to start a new vm in the blocks before hardfork,
+        // destroying the assumption that the transaction execution only uses the old vm
+        // before hardfork, leading to unexpected network splits.
+        let epoch_number = self.tx_env.epoch_number_without_proposal_window();
+        let hardfork_switch = self.consensus.hardfork_switch();
+        hardfork_switch
+            .ckb2021
+            .is_vm_version_1_and_syscalls_2_enabled(epoch_number)
+    }
+
     fn is_vm_version_2_and_syscalls_3_enabled(&self) -> bool {
         // If the proposal window is allowed to prejudge on the vm version,
         // it will cause proposal tx to start a new vm in the blocks before hardfork,
@@ -551,11 +563,18 @@ impl<DL: CellDataProvider + HeaderProvider + ExtensionProvider + Send + Sync + C
     /// Returns the version of the machine based on the script and the consensus rules.
     pub fn select_version(&self, script: &Script) -> Result<ScriptVersion, ScriptError> {
         let is_vm_version_2_and_syscalls_3_enabled = self.is_vm_version_2_and_syscalls_3_enabled();
+        let is_vm_version_1_and_syscalls_2_enabled = self.is_vm_version_1_and_syscalls_2_enabled();
         let script_hash_type = ScriptHashType::try_from(script.hash_type())
             .map_err(|err| ScriptError::InvalidScriptHashType(err.to_string()))?;
         match script_hash_type {
             ScriptHashType::Data => Ok(ScriptVersion::V0),
-            ScriptHashType::Data1 => Ok(ScriptVersion::V1),
+            ScriptHashType::Data1 => {
+                if is_vm_version_1_and_syscalls_2_enabled {
+                    Ok(ScriptVersion::V1)
+                } else {
+                    Err(ScriptError::InvalidVmVersion(1))
+                }
+            }
             ScriptHashType::Data2 => {
                 if is_vm_version_2_and_syscalls_3_enabled {
                     Ok(ScriptVersion::V2)
@@ -566,8 +585,10 @@ impl<DL: CellDataProvider + HeaderProvider + ExtensionProvider + Send + Sync + C
             ScriptHashType::Type => {
                 if is_vm_version_2_and_syscalls_3_enabled {
                     Ok(ScriptVersion::V2)
-                } else {
+                } else if is_vm_version_1_and_syscalls_2_enabled {
                     Ok(ScriptVersion::V1)
+                } else {
+                    Ok(ScriptVersion::V0)
                 }
             }
         }
