+++ spec/src/hardfork.rs
@@ -23,7 +23,7 @@ impl HardForkConfig {
         ckb2021 = self.update_2021(
             ckb2021,
             mainnet::CKB2021_START_EPOCH,
-            mainnet::RFC0028_START_EPOCH,
+            mainnet::RFC0028_RFC0032_RFC0033_RFC0034_START_EPOCH,
         )?;
 
         Ok(HardForks {
@@ -39,7 +39,7 @@ impl HardForkConfig {
         ckb2021 = self.update_2021(
             ckb2021,
             testnet::CKB2021_START_EPOCH,
-            testnet::RFC0028_START_EPOCH,
+            testnet::RFC0028_RFC0032_RFC0033_RFC0034_START_EPOCH,
         )?;
         let mut ckb2023 = CKB2023::new_builder();
         ckb2023 = self.update_2023(ckb2023, testnet::CKB2023_START_EPOCH)?;
@@ -54,14 +54,14 @@ impl HardForkConfig {
         &self,
         builder: CKB2021Builder,
         ckb2021: EpochNumber,
-        rfc_0028_start: EpochNumber,
+        rfc_0028_0032_0033_0034_start: EpochNumber,
     ) -> Result<CKB2021Builder, String> {
         let builder = builder
-            .rfc_0028(rfc_0028_start)
+            .rfc_0028(rfc_0028_0032_0033_0034_start)
             .rfc_0029(ckb2021)
             .rfc_0030(ckb2021)
             .rfc_0031(ckb2021)
-            .rfc_0032(ckb2021)
+            .rfc_0032(rfc_0028_0032_0033_0034_start)
             .rfc_0036(ckb2021)
             .rfc_0038(ckb2021);
         Ok(builder)
@@ -80,12 +80,7 @@ impl HardForkConfig {
     ///
     /// Enable features which are set to `None` at the dev default config.
     pub fn complete_with_dev_default(&self) -> Result<HardForks, String> {
-        let mut ckb2021 = CKB2021::new_builder();
-        ckb2021 = self.update_2021(
-            ckb2021,
-            testnet::CKB2021_START_EPOCH,
-            testnet::RFC0028_START_EPOCH,
-        )?;
+        let ckb2021 = CKB2021::new_dev_default();
 
         let ckb2023 = if let Some(epoch) = self.ckb2023 {
             CKB2023::new_with_specified(epoch)
@@ -93,9 +88,6 @@ impl HardForkConfig {
             CKB2023::new_dev_default()
         };
 
-        Ok(HardForks {
-            ckb2021: ckb2021.build()?,
-            ckb2023,
-        })
+        Ok(HardForks { ckb2021, ckb2023 })
     }
 }
