+++ util/constant/src/hardfork/mainnet.rs
@@ -1,8 +1,8 @@
 /// The Chain Specification name.
 pub const CHAIN_SPEC_NAME: &str = "ckb";
 
-/// hardcode rfc0028 epoch
-pub const RFC0028_START_EPOCH: u64 = 5414;
+/// hardcode rfc0028/rfc0032/rfc0033/rfc0034 epoch
+pub const RFC0028_RFC0032_RFC0033_RFC0034_START_EPOCH: u64 = 5414;
 /// First epoch number for CKB v2021, at about 2022/05/10 1:00 UTC
 // pub const CKB2021_START_EPOCH: u64 = 5414;
 pub const CKB2021_START_EPOCH: u64 = 0;
