+++ util/constant/src/hardfork/testnet.rs
@@ -1,8 +1,8 @@
 /// The Chain Specification name.
 pub const CHAIN_SPEC_NAME: &str = "ckb_testnet";
 
-/// hardcode rfc0028 epoch
-pub const RFC0028_START_EPOCH: u64 = 3113;
+/// hardcode rfc0028/rfc0032/rfc0033/rfc0034 epoch
+pub const RFC0028_RFC0032_RFC0033_RFC0034_START_EPOCH: u64 = 3113;
 /// First epoch number for CKB v2021, at about 2021/10/24 3:15 UTC.
 // pub const CKB2021_START_EPOCH: u64 = 3113;
 pub const CKB2021_START_EPOCH: u64 = 0;
